requirements.txt
addict==2.4.0
axial_positional_embedding==0.2.1
CoLT5-attention==0.11.1
contourpy==1.3.0
cycler==0.12.1
einops==0.8.0
fonttools==4.53.1
fsspec==2024.6.1
fvcore==0.1.5.post20221221
huggingface-hub==0.24.6
importlib_metadata==8.4.0
iopath==0.1.10
kiwisolver==1.4.5
linear-attention-transformer==0.19.1
linformer==0.2.3
local-attention==1.9.15
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
mkl-service==2.4.0
mmcls==0.25.0
mmcv-full==1.7.2
mmengine==0.10.5
MultiScaleDeformableAttention==1.0
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
pillow
platformdirs==4.2.2
portalocker==2.10.1
prettytable==3.11.0
product_key_memory==0.2.11
Pygments==2.18.0
pyparsing==3.1.4
PySocks
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML
requests
rich==13.9.4
seaborn==0.13.2
six==1.16.0
sympy
tabulate==0.9.0
termcolor==2.4.0
thop==0.1.1.post2209072238
timm==0.6.12
tomli==2.0.1
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.5
triton==2.1.0
typing_extensions
tzdata==2024.2
urllib3
wcwidth==0.2.13
yacs==0.1.8
yapf==0.40.2
zipp==3.20.1
